pyproject.toml
[project]
name = "qa-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
	"alembic>=1.13.0",
	"email-validator>=2.2.0",
	"faiss-cpu>=1.8.0",
	"fastapi>=0.115.0",
	"httpx>=0.27.0",
	"langchain-community>=0.2.0",
	"langchain-core>=0.2.0",
	"langchain-openai>=0.1.0",
	"langchain-text-splitters>=0.2.0",
	"loguru>=0.7.0",
	"openai>=1.35.0",
	"psycopg2-binary>=2.9.0",
	"pydantic>=2.8.0",
	"pypdf>=4.0.0",
	"python-dotenv>=1.0.0",
	"python-multipart>=0.0.9",
	"sentence-transformers>=2.7.0",
	"SQLAlchemy>=2.0.0",
	"tenacity>=8.4.0",
	"tiktoken>=0.7.0",
	"uvicorn[standard]>=0.30.0",
]
